refactor edit screen into reusable component
