+++ app/frontend/components/domains/requirement-template/screens/base-edit-requirement-template-screen/sections-display.tsx
@@ -6,13 +6,13 @@ import React, { useEffect } from "react"
 import { useFieldArray, useFormContext } from "react-hook-form"
 import { useTranslation } from "react-i18next"
 import { v4 as uuidv4 } from "uuid"
+import { IRequirementTemplateForm, formScrollToId } from "."
 import { useMst } from "../../../../../setup/root"
 import { IRequirementTemplateSectionAttributes } from "../../../../../types/api-request"
 import { EditableInputWithControls } from "../../../../shared/editable-input-with-controls"
 import { RemoveConfirmationModal } from "../../../../shared/modals/remove-confirmation-modal"
 import { RequirementBlockAccordion } from "../../../requirements-library/requirement-block-accordion"
 import { RequirementsLibraryDrawer } from "../../../requirements-library/requirements-library-drawer"
-import { IRequirementTemplateForm, formScrollToId } from "./index"
 
 interface IProps {
   isCollapsedAll?: boolean
