+++ app/frontend/components/domains/requirement-template/screens/base-edit-requirement-template-screen/sections-dnd/droppable-section.tsx
