+++ app/frontend/i18n/i18n.ts
@@ -234,6 +234,9 @@ const options = {
           markedForRemoval: 'Click "Save changes" to confirm removal',
           proceed: "Proceed",
           copyNoun: "Copy",
+          options: "Options",
+          na: "N/A",
+          share: "Share",
         },
         notification: {
           title: "Notifications",
@@ -1418,6 +1421,11 @@ const options = {
             startWithBlank: "Start with blank permit",
             copyFromLive: "Copy from live permit",
             copyFromThis: "Copy from this",
+          },
+          edit: {
+            lastFetched: "Last fetched",
+            fetchLatest: "Fetch latest",
+            auditLog: "Audit log",
           },
           fields: {
             nickname: "Nickname",
