+++ app/models/early_access_requirement_template.rb
@@ -1,9 +1,13 @@
 class EarlyAccessRequirementTemplate < RequirementTemplate
-  # has_one :template_version, dependent: :destroy, foreign_key: :requirement_template_id
+  has_one :template_version, dependent: :destroy, foreign_key: :requirement_template_id
 
   before_validation :ensure_one_published_template_version
 
   validate :has_one_published_template_version
+
+  def label
+    nickname || "asdf"
+  end
 
   private
